\name{imKmeans-package}
\alias{imKmeans-package}
\alias{imKmeans}
\docType{package}
\title{
\packageTitle{imKmeans}
}
\description{
\packageDescription{imKmeans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imKmeans}
\packageIndices{imKmeans}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{imKmeans}

Maintainer: \packageMaintainer{imKmeans}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
