% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imKmeans-package.R, R/imKmeans.R
\docType{package}
\name{imKmeans}
\alias{imKmeans}
\alias{imKmeans-package}
\title{imKmeans package}
\usage{
imKmeans(dat, cl)
}
\arguments{
\item{dat}{data as matrix}

\item{cl}{number of clusters}
}
\value{
cluster cluster that assigned for each obs.

aic AIC

bic BIC
}
\description{
packages only for improved k-means method

improved k-means function
}
\examples{
imKmeans(iris[1:4],3)
}

