\name{imKmeans-package}
\alias{imKmeans-package}
\docType{package}
\title{
\packageTitle{imKmeans}
}
\description{
\packageDescription{imKmeans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imKmeans}
\packageIndices{imKmeans}
}
\author{
\packageAuthor{imKmeans}

Maintainer: \packageMaintainer{imKmeans}
}
\references{
Toyoda H. and Ikehara K. 2011 An improved method using k-means to determine the optimal number of clusters, considering the relations between several variables. The Japanese Journal of Psychology,Vol.82, No.1, pp.32-40.
}
\examples{
imKmeans(iris[1:4],3)
}
